refactor: move diff protocol to messenger
