+++ extension/src/domain/diff/types.ts
@@ -75,52 +75,3 @@ export type DiffTarget =
   | { kind: "pull"; prNumber: number }
   | { kind: "commit"; sha: string }
   | { kind: "compare"; base: string; head: string };
-
-// content ↔ background messages (chrome.runtime only serializes JSON)
-export type SemanticDiffRequest = {
-  type: "semanticDiff";
-  owner: string;
-  repo: string;
-  target: DiffTarget;
-  path: string;
-  force?: boolean; // render past the 25MB guard ("Render anyway" click)
-};
-
-export type PrefetchRequest = { type: "prefetch"; owner: string; repo: string; prNumber: number };
-export type BackgroundRequest = SemanticDiffRequest | PrefetchRequest;
-
-export type BackgroundError =
-  | "access-token-missing"
-  | "auth-failed"
-  | "rate-limited"
-  | "fetch-failed"
-  | "diff-failed"
-  | "not-unity-yaml";
-
-// Errors that a completed sign-in can clear. Everything else renders a plain error card.
-export type AuthError = Extract<BackgroundError, "access-token-missing" | "auth-failed">;
-
-export function isAuthError(error: BackgroundError): error is AuthError {
-  return error === "access-token-missing" || error === "auth-failed";
-}
-
-export type SemanticDiffResponse =
-  | { ok: true; json: DiffV2; pending?: boolean }
-  | { ok: false; error: BackgroundError }
-  | { ok: false; error: "too-large"; bytes: number };
-
-// Background resolution outcome. Anything but "complete" means a manual retry can help.
-export type ResolutionStatus = "complete" | "rateLimited" | "failed";
-
-// Async push background → content (stage 2 of the two-stage response).
-export type GuidResolvedPush = {
-  type: "guidResolved";
-  owner: string;
-  repo: string;
-  target: DiffTarget;
-  path: string;
-  resolved: Record<string, string>;
-  json?: DiffV2; // The final push can carry a restructured diff after updateSources.
-  done: boolean; // Resolution finished. Even with empty resolved, done stops the indicator.
-  status?: ResolutionStatus; // On every done push. Keep the indicator unless "complete".
-};
